src/middleware/TokenMiddleware.ts
import { Request, Response, NextFunction } from "express";
import { db } from "../data/database";

const verifyToken = (req: Request, res: Response, next: NextFunction): void => {
  const token = req.headers.authorization?.split(" ")[1];
  if (!token) {
    res.status(401).json({ error: "Unauthorized" });
    return;
  }
  const user = db.get(token);
  if (!user) {
    res.status(401).json({ error: "Unauthorized" });
    return;
  }
  (req as any).user = user;
  (req as any).token = token;
  next();
};

export { verifyToken };
